+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - New `is` operator added to `text` condition.
 - New config unit test condition `content_matches`.
 - Field `init_values` added to the `memory` cache.
+- New `split` operator added to `json` processor.
 
 ## 3.1.1 - 2019-09-23